+++ HISTORY.md
@@ -12,6 +12,7 @@ unreleased
   * remove unnecessary devDependency `readable-stream`
   * refactor: use object spread to copy error headers
   * refactor: use replaceAll instead of replace with a regex
+  * refactor: replace setHeaders function with optimized inline header setting
 
 v2.0.0 / 2024-09-02
 ==================